+++ src/lib.rs
@@ -4,9 +4,9 @@
 //!
 //! This crate is primarily designed to be distributed as a standalone binary. And has no intention to maintain as a library dependency for other projects for now.
 
-mod client;
 mod config;
 mod document;
+mod editor_info;
 mod error;
 mod fs;
 mod logger;
@@ -14,6 +14,7 @@ mod parser;
 mod providers;
 mod resolver;
 mod server;
+mod server_info;
 #[doc(hidden)]
 pub use crate::server::PathServer;
 #[doc(hidden)]